Refactor WatcherFactory to ensure fetch is always called
Every started instance of PersistentWatcher must re-fetch its
configured path when the curator is replaced. Previously, this
mostly happened by passing a runnable to the replaceCurator
method. If there are several watchers and they all call
replaceCurator on their parent however, only one will have
fetch called.

Instead, we track created PersistentWatchers and instruct
each to fetch when the curator is replaced. This ensures
we always subscribe to new change events and simplifies
the code.